end of 5-13

+++ src/AuctionService/Controllers/AuctionsController.cs
@@ -6,6 +6,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -26,24 +27,24 @@ public class AuctionsController : ControllerBase
         _publishEndpoint = publishEndpoint;
     }
 
-   /*  [HttpGet]
-    public async Task<ActionResult<List<AuctionDto>>> GetAllAuctions()
-    {
-        var auctions = await _context.Auctions
-            .Include(x=> x.Item)
-            .OrderBy(x=> x.Item.Make)
-            .ToListAsync();
+    /*  [HttpGet]
+     public async Task<ActionResult<List<AuctionDto>>> GetAllAuctions()
+     {
+         var auctions = await _context.Auctions
+             .Include(x=> x.Item)
+             .OrderBy(x=> x.Item.Make)
+             .ToListAsync();
 
-        return _mapper.Map<List<AuctionDto>>(auctions);
-    } */
+         return _mapper.Map<List<AuctionDto>>(auctions);
+     } */
 
-     [HttpGet]
+    [HttpGet]
     public async Task<ActionResult<List<AuctionDto>>> GetAllAuctions(string date)
     {
-        var query = _context.Auctions.OrderBy(x=> x.Item.Make).AsQueryable();
-        if(!string.IsNullOrEmpty(date))
+        var query = _context.Auctions.OrderBy(x => x.Item.Make).AsQueryable();
+        if (!string.IsNullOrEmpty(date))
         {
-            query = query.Where(x=> x.UpdatedAt.CompareTo(DateTime.Parse(date).ToUniversalTime())>0);
+            query = query.Where(x => x.UpdatedAt.CompareTo(DateTime.Parse(date).ToUniversalTime()) > 0);
         }
         return await query.ProjectTo<AuctionDto>(_mapper.ConfigurationProvider).ToListAsync();
     }
@@ -52,43 +53,47 @@ public class AuctionsController : ControllerBase
     public async Task<ActionResult<AuctionDto>> GetAuctionById(Guid id)
     {
         var auction = await _context.Auctions
-            .Include(x=> x.Item)
-            .FirstOrDefaultAsync(x=> x.Id == id);
+            .Include(x => x.Item)
+            .FirstOrDefaultAsync(x => x.Id == id);
 
-        if(auction == null) return NotFound();    
+        if (auction == null) return NotFound();
 
-        return _mapper.Map<AuctionDto>(auction);  
+        return _mapper.Map<AuctionDto>(auction);
     }
 
+    [Authorize]
     [HttpPost]
     public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto auctionDto)
     {
         var auction = _mapper.Map<Auction>(auctionDto);
         //TODO: add current user as seller
-        auction.Seller = "test";
+        // auction.Seller = "test";
+        auction.Seller = User.Identity.Name;
 
         _context.Auctions.Add(auction);
 
-         var newAuction = _mapper.Map<AuctionDto>(auction);
+        var newAuction = _mapper.Map<AuctionDto>(auction);
 
         await _publishEndpoint.Publish(_mapper.Map<AuctionCreated>(newAuction));
 
-        var result = await _context.SaveChangesAsync()> 0;
+        var result = await _context.SaveChangesAsync() > 0;
 
-        if(!result) return BadRequest("Could not save changes to the DB");
+        if (!result) return BadRequest("Could not save changes to the DB");
 
 
-        return CreatedAtAction(nameof(GetAuctionById), new {auction.Id}, newAuction);
+        return CreatedAtAction(nameof(GetAuctionById), new { auction.Id }, newAuction);
     }
 
+    [Authorize]
     [HttpPut("{id}")]
     public async Task<ActionResult> UpdateAuction(Guid id, UpdateAuctionDto updateAuctionDto)
     {
-        var auction = await _context.Auctions.Include(x=> x.Item).FirstOrDefaultAsync(x=> x.Id == id);
+        var auction = await _context.Auctions.Include(x => x.Item).FirstOrDefaultAsync(x => x.Id == id);
 
-        if(auction == null) return NotFound();
+        if (auction == null) return NotFound();
 
         //TODO: check seller == username
+        if (auction.Seller != User.Identity.Name) return Forbid();
 
         auction.Item.Make = updateAuctionDto.Make ?? auction.Item.Make;
         auction.Item.Model = updateAuctionDto.Model ?? auction.Item.Model;
@@ -98,29 +103,31 @@ public class AuctionsController : ControllerBase
 
         await _publishEndpoint.Publish(_mapper.Map<AuctionUpdated>(auction));
 
-        var result = await _context.SaveChangesAsync() > 0 ;
+        var result = await _context.SaveChangesAsync() > 0;
 
-        if(result) return Ok();
+        if (result) return Ok();
 
         return BadRequest("Problem saving changes");
     }
 
+    [Authorize]
     [HttpDelete("{id}")]
     public async Task<ActionResult> DeleteAuction(Guid id)
     {
         var auction = await _context.Auctions.FindAsync(id);
 
-        if(auction == null) return NotFound();
+        if (auction == null) return NotFound();
 
         //TODO: check seller == username
+        if (auction.Seller != User.Identity.Name) return Forbid();
 
         _context.Auctions.Remove(auction);
 
         await _publishEndpoint.Publish<AuctionDeleted>(new { Id = auction.Id.ToString() });
 
-        var result = await _context.SaveChangesAsync()>0;
+        var result = await _context.SaveChangesAsync() > 0;
 
-        if(!result) return BadRequest("Could not update DB");
+        if (!result) return BadRequest("Could not update DB");
 
         return Ok();
     }
